+++ backend/app/models/supply_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Float
+from sqlalchemy import Column, Integer, ForeignKey, Float, Numeric
 from sqlalchemy.orm import relationship
 from app.core.database import Base
 
@@ -9,8 +9,8 @@ class SupplyItem(Base):
     supply_id = Column(Integer, ForeignKey("supplies.id"), nullable=False)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
     quantity = Column(Integer, nullable=False)
-    cost_price = Column(Float, nullable=True)     # Себестоимость
-    unit_price = Column(Float, nullable=True)     # Закупочная цена (если отличается)
+    cost_price = Column(Numeric(12, 2), nullable=True)     # Себестоимость
+    unit_price = Column(Numeric(12, 2), nullable=True)     # Закупочная цена (если отличается)
 
     supply = relationship("Supply", back_populates="items")
     product = relationship("Product")
